add crypto guard app class
