+++ src/main.cpp
@@ -1,103 +1,12 @@
-#include "cmd_options.h"
-#include "crypto_guard_ctx.h"
-#include <array>
+#include "crypto_guard_app.h"
 #include <iostream>
-#include <openssl/evp.h>
 #include <print>
 #include <stdexcept>
-#include <string>
-
-struct AesCipherParams {
-    static const size_t KEY_SIZE = 32;             // AES-256 key size
-    static const size_t IV_SIZE = 16;              // AES block size (IV length)
-    const EVP_CIPHER *cipher = EVP_aes_256_cbc();  // Cipher algorithm
-
-    int encrypt;                              // 1 for encryption, 0 for decryption
-    std::array<unsigned char, KEY_SIZE> key;  // Encryption key
-    std::array<unsigned char, IV_SIZE> iv;    // Initialization vector
-};
-
-AesCipherParams CreateChiperParamsFromPassword(std::string_view password) {
-    AesCipherParams params;
-    constexpr std::array<unsigned char, 8> salt = {'1', '2', '3', '4', '5', '6', '7', '8'};
-
-    int result = EVP_BytesToKey(params.cipher, EVP_sha256(), salt.data(),
-                                reinterpret_cast<const unsigned char *>(password.data()), password.size(), 1,
-                                params.key.data(), params.iv.data());
-
-    if (result == 0) {
-        throw std::runtime_error{"Failed to create a key from password"};
-    }
-
-    return params;
-}
 
 int main(int argc, char *argv[]) {
     try {
-        //
-        // OpenSSL пример использования:
-        //
-        std::string input = "01234567890123456789";
-        std::string output;
-
-        OpenSSL_add_all_algorithms();
-
-        auto params = CreateChiperParamsFromPassword("12341234");
-        params.encrypt = 1;
-        auto *ctx = EVP_CIPHER_CTX_new();
-
-        // Инициализируем cipher
-        EVP_CipherInit_ex(ctx, params.cipher, nullptr, params.key.data(), params.iv.data(), params.encrypt);
-
-        std::vector<unsigned char> outBuf(16 + EVP_MAX_BLOCK_LENGTH);
-        std::vector<unsigned char> inBuf(16);
-        int outLen;
-
-        // Обрабатываем первые N символов
-        EVP_CipherUpdate(ctx, outBuf.data(), &outLen, inBuf.data(), static_cast<int>(16));
-        for (int i = 0; i < outLen; ++i) {
-            output.push_back(outBuf[i]);
-        }
-
-        // Обрабатываем оставшиеся символы
-        EVP_CipherUpdate(ctx, outBuf.data(), &outLen, inBuf.data(), static_cast<int>(16));
-        for (int i = 0; i < outLen; ++i) {
-            output.push_back(outBuf[i]);
-        }
-
-        // Заканчиваем работу с cipher
-        EVP_CipherFinal_ex(ctx, outBuf.data(), &outLen);
-        for (int i = 0; i < outLen; ++i) {
-            output.push_back(outBuf[i]);
-        }
-        EVP_CIPHER_CTX_free(ctx);
-        std::print("String encoded successfully. Result: '{}'\n\n", output);
-        EVP_cleanup();
-        //
-        // Конец примера
-        //
-
-        CryptoGuard::ProgramOptions options;
-
-        CryptoGuard::CryptoGuardCtx cryptoCtx;
-
-        using COMMAND_TYPE = CryptoGuard::ProgramOptions::COMMAND_TYPE;
-        switch (options.GetCommand()) {
-        case COMMAND_TYPE::ENCRYPT:
-            std::print("File encoded successfully\n");
-            break;
-
-        case COMMAND_TYPE::DECRYPT:
-            std::print("File decoded successfully\n");
-            break;
-
-        case COMMAND_TYPE::CHECKSUM:
-            std::print("Checksum: {}\n", "CHECKSUM_NOT_IMPLEMENTED");
-            break;
-
-        default:
-            throw std::runtime_error{"Unsupported command"};
-        }
+        CryptoGuard::CryptoGuardApp app(argc, argv);
+        app.Run();
 
     } catch (const std::exception &e) {
         std::print(std::cerr, "Error: {}\n", e.what());
